:recycle: refactor add works to collections job
all errors to be captured by good job
